Veja a descrição
- [Simulador] Desabilita loop do time que está desabilitado e aumenta a velocidade dos robôs
- [Helpers] Cria helpers/decorators para calcular tempo de execução
- [Estratégia] Calcula velocidade da bola 
- [Geral]  Correções de padronização
