+++ hades.py
@@ -268,19 +268,3 @@ class Hades(QThread):
 
     def eventsendMessage(self, robotId, message):
         self.hermes.sendMessage(robotId, message)
-
-
-def timeToFinish(method):
-    # This decorator returns time elapsed on execution of a method
-    # HOW TO USE
-    # Before the method, place @timeToFinish
-    # In the terminal will be printed the time elapsed on method execution
-
-    def timed(*args, **kwargs):
-        tStart = time.time()
-        result = method(*args, **kwargs)
-        tEnd = time.time()
-
-        print("{:.3f} sec".format(tEnd-tStart))
-        return result
-    return timed
